Make list workspaces works
