--- components/ide/jetbrains/toolbox/src/main/kotlin/dto.kt
@@ -1,14 +1,0 @@
-package toolbox.gateway.sample
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class EnvironmentDTO(
-    val id: String,
-    val name: String,
-)
-
-@Serializable
-data class EnvironmentsDTO(
-    val environments: List<EnvironmentDTO>
-)
