+++ components/ide/jetbrains/toolbox/src/main/kotlin/io/gitpod/toolbox/gateway/GitpodGatewayExtension.kt
@@ -1,0 +1,16 @@
+package io.gitpod.toolbox.gateway
+
+import com.jetbrains.toolbox.gateway.GatewayExtension
+import com.jetbrains.toolbox.gateway.RemoteEnvironmentConsumer
+import com.jetbrains.toolbox.gateway.RemoteProvider
+import com.jetbrains.toolbox.gateway.ToolboxServiceLocator
+import kotlinx.coroutines.CoroutineScope
+
+class GitpodGatewayExtension : GatewayExtension {
+    override fun createRemoteProviderPluginInstance(serviceLocator: ToolboxServiceLocator): RemoteProvider {
+        return GitpodRemoteProvider(
+            serviceLocator.getService(RemoteEnvironmentConsumer::class.java),
+            serviceLocator.getService(CoroutineScope::class.java),
+        )
+    }
+}
