+++ components/ide/jetbrains/toolbox/src/main/resources/META-INF/services/com.jetbrains.toolbox.gateway.GatewayExtension
@@ -1,1 +1,1 @@
-toolbox.gateway.sample.SampleGatewayExtension
+io.gitpod.toolbox.gateway.GitpodGatewayExtension
